feat: fix/test all filters
