+++ src/webapp/reports/csy-audit-trauma/AuditViewModel.ts
@@ -1,5 +1,5 @@
 import { Config } from "../../../domain/common/entities/Config";
-import { AuditItem } from "../../../domain/reports/csy-audit-trauma/entities/AuditItem";
+import { AuditItem } from "../../../domain/reports/csy-audit-emergency/entities/AuditItem";
 
 export interface AuditViewModel {
     id: string;
